Compatibility with Magento 2.3.x using integrated amqp/RABBITMQ
