--- etc/ce_mq.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Rcason_Mq:etc/ce_mq.xsd">
-    <ceQueue name="lcp.product.prewarm" broker="mysql"
-             messageSchema="int"
-             consumerInterface="Elgentos\LargeConfigProducts\Model\Consumer"/>
-</config>
