+++ etc/queue_consumer.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework-message-queue:etc/queue_consumer.xsd">
+    <!-- name: The name of the consumer -->
+    <!-- queue: Defines the queue name to send the message to -->
+    <!-- connection: For AMQP connections, the connection name must match the connection attribute in the queue_topology.xml file. Otherwise, the connection name must be db -->
+    <!-- consumerInstance: The Magento class name that consumes the message -->
+    <consumer name="elgentos_magento_lcp_product_prewarm" queue="elgentos_magento_lcp_product_prewarm" connection="amqp" handler="Elgentos\LargeConfigProducts\Model\MessageQueues\Consumer::processMessage"/>
+</config>
