+++ etc/queue_publisher.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework-message-queue:etc/publisher.xsd">
+    <publisher topic="elgentos.magento.lcp.product.prewarm">
+        <connection name="amqp" exchange="elgentos.magento.lcp.product.prewarm" />
+    </publisher>
+</config>
